actualizacion de la rama jose camargo

+++ routes/web.php
@@ -24,26 +24,26 @@ Route::middleware([
 
  
 // Ruta para mostrar la lista de usuarios y manejar la búsqueda
-Route::get('/user-management', [UserManagementController::class, 'index'])->name('user-management.index');
+Route::get('/user-management', [UserManagementController::class, 'index'])->name('user-management.index')->middleware('redirect.not.admin');
 
 // Ruta para actualizar la contraseña de un usuario
-Route::post('/user-management/update/{id}', [UserManagementController::class, 'update'])->name('user-management.update');
+Route::post('/user-management/update/{id}', [UserManagementController::class, 'update'])->name('user-management.update')->middleware('redirect.not.admin');
 
 // Ruta para actualizar los roles de un usuario
-Route::post('/user-management/update-roles/{id}', [UserManagementController::class, 'updateRoles'])->name('user-management.update-roles');
+Route::post('/user-management/update-roles/{id}', [UserManagementController::class, 'updateRoles'])->name('user-management.update-roles')->middleware('redirect.not.admin');
 
 // Ruta para la búsqueda de usuarios
-Route::get('/user-management/search', [UserManagementController::class, 'search'])->name('user-management.search');
+Route::get('/user-management/search', [UserManagementController::class, 'search'])->name('user-management.search')->middleware('redirect.not.admin');
 
 
 // productos 
-Route::get('/productos', [ProductoController::class, 'index'])->name('productos.index');
-Route::get('/productos/create', [ProductoController::class, 'create'])->name('productos.create');
+Route::get('/productos', [ProductoController::class, 'index'])->name('productos.index')->middleware('redirect.not.admin');
+Route::get('/productos/create', [ProductoController::class, 'create'])->name('productos.create')->middleware('redirect.not.admin');
 Route::post('/productos', [ProductoController::class, 'store'])->name('productos.store');
-Route::get('/productos/{producto}', [ProductoController::class, 'show'])->name('productos.show');
-Route::get('/productos/{producto}/edit', [ProductoController::class, 'edit'])->name('productos.edit');
-Route::put('/productos/{producto}', [ProductoController::class, 'update'])->name('productos.update');
-Route::delete('/productos/{producto}', [ProductoController::class, 'destroy'])->name('productos.destroy');
+Route::get('/productos/{producto}', [ProductoController::class, 'show'])->name('productos.show')->middleware('redirect.not.admin');
+Route::get('/productos/{producto}/edit', [ProductoController::class, 'edit'])->name('productos.edit')->middleware('redirect.not.admin');
+Route::put('/productos/{producto}', [ProductoController::class, 'update'])->name('productos.update')->middleware('redirect.not.admin');
+Route::delete('/productos/{producto}', [ProductoController::class, 'destroy'])->name('productos.destroy')->middleware('redirect.not.admin');
 
 
 // Rutas para el carrito
